Purchase Agreements Fields
- Move fields it own class
- Add "Display Title" field
